fixed64: hand-fix batch 2 -- drive strict errors to zero
Resolve the final 34 strict-type errors so a -DZX_STRONG_FIXED build is
clean across the whole tree:

- double<->fixed geometry/BSP math: wrap the fixed operand in double()/
  float() so nodebuild, p_slopes, gl_decal, gl_portaldata, p_glnodes,
  a_flechette and the decallib clamp stay pure floating-point (FRACUNIT
  is now a fixed_t, which is what pulled doubles into fixed overloads).
- function args: cast to the parameter type -- SoundPoint y/z (LONG),
  P_LineAttack pitch and ColorMatcher.Pick g/b (int), BOTS path test.
- angle_t reinterpret hazards: read a wrapping BAM value back as a signed
  int32 before it becomes a fixed_t operand. Factor that reinterpret into
  zx::AngleAsSignedFixed and reuse it in the security-camera swing, the
  player turn delta, and the GL portal view interpolation (which now
  shares zx::InterpolateAngleBAM with R_InterpolateView). Add regression
  tests pinning the >=180-degree and right-turn-delta cases.
- unsigned->fixed sinks (v_draw errorAdj, sbarinfo clip), the DScroller
  ctor ambiguity, and an over-nested (int) wrap in p_acs.

Behavior-preserving: every cast reproduces the original 32-bit result,
and the flag-off build stays byte-identical.

+++ src/zandronum/src/g_doom/a_bossbrain.cpp
@@ -172,7 +172,7 @@ DEFINE_ACTION_FUNCTION_PARAMS(AActor, A_BrainSpit)
 
 			// [BC] If we're the server, tell clients create the sound.
 			if ( NETWORK_GetState( ) == NETSTATE_SERVER )
-				SERVERCOMMANDS_SoundPoint((LONG)( self->x), self->y, self->z, CHAN_WEAPON, self->AttackSound, 1, ATTN_NONE );
+				SERVERCOMMANDS_SoundPoint((LONG)( self->x), (LONG)self->y, (LONG)self->z, CHAN_WEAPON, self->AttackSound, 1, ATTN_NONE );
 		}
 		else
 		{
@@ -181,7 +181,7 @@ DEFINE_ACTION_FUNCTION_PARAMS(AActor, A_BrainSpit)
 
 			// [BC] If we're the server, tell clients create the sound.
 			if ( NETWORK_GetState( ) == NETSTATE_SERVER )
-				SERVERCOMMANDS_SoundPoint((LONG)( self->x), self->y, self->z, CHAN_WEAPON, "brain/spit", 1, ATTN_NONE );
+				SERVERCOMMANDS_SoundPoint((LONG)( self->x), (LONG)self->y, (LONG)self->z, CHAN_WEAPON, "brain/spit", 1, ATTN_NONE );
 		}
 	}
 }
@@ -222,7 +222,7 @@ static void SpawnFly(AActor *self, const PClass *spawntype, FSoundID sound)
 	if (( NETWORK_GetState( ) == NETSTATE_SERVER ) && ( fog ))
 	{
 		SERVERCOMMANDS_SpawnThing( fog );
-		SERVERCOMMANDS_SoundPoint((LONG)( fog->x), fog->y, fog->z, CHAN_BODY, S_GetName( sound ), 1, ATTN_NORM );
+		SERVERCOMMANDS_SoundPoint((LONG)( fog->x), (LONG)fog->y, (LONG)fog->z, CHAN_BODY, S_GetName( sound ), 1, ATTN_NORM );
 	}
 
 	FName SpawnName;
